modules/utility/cmd/requirements.txt
pyDataCanvas >= 0.4.0,<0.5.0